+++ bip_peer/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name          = "bip_peer"
-version       = "0.3.1"
+version       = "0.4.0"
 description   = "Communication with bittorrent peers via peer wire protocol"
 
 authors       = ["Andrew <amiller4421@gmail.com>"]
@@ -14,12 +14,12 @@ keywords      = ["peer", "wire", "protocol", "pwp", "bittorrent"]
 license       = "MIT/Apache-2.0"
 
 [dependencies]
-bip_bencode   = "0.3"
+bip_bencode   = "0.4"
 bip_util      = "0.5"
 bytes         = "0.4"
 byteorder     = "1.0"
-crossbeam     = "0.2"
-error-chain   = "0.9"
+crossbeam     = "0.3"
+error-chain   = "0.11"
 futures       = "0.1"
 tokio-core    = "0.1"
 tokio-io      = "0.1"
