+++ bip_peer/src/message/bencode.rs
@@ -6,7 +6,7 @@ use std::str;
 use bip_bencode::{BencodeConvertError, BConvert, BRefAccess, BDictAccess};
 use bip_util::convert;
 
-use message::bits_extension::ExtendedType;
+use message::bits_ext::ExtendedType;
 
 pub const CONVERT: IoErrorBencodeConvert = IoErrorBencodeConvert;
 
@@ -31,13 +31,13 @@ pub const CLIENT_IPV4_ADDR_KEY:    &'static [u8] = b"ipv4";
 pub const CLIENT_MAX_REQUESTS_KEY: &'static [u8] = b"reqq";
 pub const METADATA_SIZE_KEY:       &'static [u8] = b"metadata_size";
 
-pub fn parse_id_map<'a, B>(root: &BDictAccess<'a, B>) -> HashMap<ExtendedType, u8>
-    where B: BRefAccess<'a> {
+pub fn parse_id_map<K, V>(root: &BDictAccess<K, V>) -> HashMap<ExtendedType, u8>
+    where V: BRefAccess, V::BKey: AsRef<[u8]> {
     let mut id_map = HashMap::new();
     
     if let Ok(ben_id_map) = CONVERT.lookup_and_convert_dict(root, ID_MAP_KEY) {
         for (id, ben_value) in ben_id_map.to_list() {
-            match (str::from_utf8(id), CONVERT.convert_int(ben_value, id)) {
+            match (str::from_utf8(id.as_ref()), CONVERT.convert_int(ben_value, id)) {
                 (Ok(str_id), Ok(value)) => { id_map.insert(ExtendedType::from_id(str_id), value as u8); },
                 _                       => ()
             }
@@ -47,15 +47,15 @@ pub fn parse_id_map<'a, B>(root: &BDictAccess<'a, B>) -> HashMap<ExtendedType, u
     id_map
 }
 
-pub fn parse_client_id<'a, B>(root: &BDictAccess<'a, B>) -> Option<String>
-    where B: BRefAccess<'a> {
+pub fn parse_client_id<K, V>(root: &BDictAccess<K, V>) -> Option<String>
+    where V: BRefAccess {
     CONVERT.lookup_and_convert_str(root, CLIENT_ID_KEY)
         .map(|id| id.to_string())
         .ok()
 }
 
-pub fn parse_client_tcp_port<'a, B>(root: &BDictAccess<'a, B>) -> Option<u16>
-    where B: BRefAccess<'a> {
+pub fn parse_client_tcp_port<K, V>(root: &BDictAccess<K, V>) -> Option<u16>
+    where V: BRefAccess {
     CONVERT.lookup_and_convert_int(root, CLIENT_TCP_PORT_KEY)
         .ok()
         .and_then(|port| {
@@ -67,8 +67,8 @@ pub fn parse_client_tcp_port<'a, B>(root: &BDictAccess<'a, B>) -> Option<u16>
         })
 }
 
-pub fn parse_our_ip<'a, B>(root: &BDictAccess<'a, B>) -> Option<IpAddr>
-    where B: BRefAccess<'a> {
+pub fn parse_our_ip<K, V>(root: &BDictAccess<K, V>) -> Option<IpAddr>
+    where V: BRefAccess {
     CONVERT.lookup_and_convert_bytes(root, OUR_IP_KEY)
         .ok()
         .and_then(|ip_bytes| {
@@ -82,8 +82,8 @@ pub fn parse_our_ip<'a, B>(root: &BDictAccess<'a, B>) -> Option<IpAddr>
         })
 }
 
-pub fn parse_client_ipv6_addr<'a, B>(root: &BDictAccess<'a, B>) -> Option<Ipv6Addr>
-    where B: BRefAccess<'a> {
+pub fn parse_client_ipv6_addr<K, V>(root: &BDictAccess<K, V>) -> Option<Ipv6Addr>
+    where V: BRefAccess {
     CONVERT.lookup_and_convert_bytes(root, CLIENT_IPV6_ADDR_KEY)
         .ok()
         .and_then(|ipv6_bytes| {
@@ -95,8 +95,8 @@ pub fn parse_client_ipv6_addr<'a, B>(root: &BDictAccess<'a, B>) -> Option<Ipv6Ad
         })
 }
 
-pub fn parse_client_ipv4_addr<'a, B>(root: &BDictAccess<'a, B>) -> Option<Ipv4Addr> 
-    where B: BRefAccess<'a> {
+pub fn parse_client_ipv4_addr<K, V>(root: &BDictAccess<K, V>) -> Option<Ipv4Addr> 
+    where V: BRefAccess {
     CONVERT.lookup_and_convert_bytes(root, CLIENT_IPV4_ADDR_KEY)
         .ok()
         .and_then(|ipv4_bytes| {
@@ -108,14 +108,14 @@ pub fn parse_client_ipv4_addr<'a, B>(root: &BDictAccess<'a, B>) -> Option<Ipv4Ad
         })
 }
 
-pub fn parse_client_max_requests<'a, B>(root: &BDictAccess<'a, B>) -> Option<i64>
-    where B: BRefAccess<'a> {
+pub fn parse_client_max_requests<K, V>(root: &BDictAccess<K, V>) -> Option<i64>
+    where V: BRefAccess {
     CONVERT.lookup_and_convert_int(root, CLIENT_MAX_REQUESTS_KEY)
         .ok()
 }
 
-pub fn parse_metadata_size<'a, B>(root: &BDictAccess<'a, B>) -> Option<i64>
-    where B: BRefAccess<'a> {
+pub fn parse_metadata_size<K, V>(root: &BDictAccess<K, V>) -> Option<i64>
+    where V: BRefAccess {
     CONVERT.lookup_and_convert_int(root, METADATA_SIZE_KEY)
         .ok()
 }
@@ -137,17 +137,17 @@ pub const MESSAGE_TYPE_KEY: &'static [u8] = b"msg_type";
 pub const PIECE_INDEX_KEY:  &'static [u8] = b"piece";
 pub const TOTAL_SIZE_KEY:   &'static [u8] = b"total_size";
 
-pub fn parse_message_type<'a, B>(root: &BDictAccess<'a, B>) -> io::Result<u8>
-    where B: BRefAccess<'a> {
+pub fn parse_message_type<K, V>(root: &BDictAccess<K, V>) -> io::Result<u8>
+    where V: BRefAccess {
     CONVERT.lookup_and_convert_int(root, MESSAGE_TYPE_KEY).map(|msg_type| msg_type as u8).into()
 }
 
-pub fn parse_piece_index<'a, B>(root: &BDictAccess<'a, B>) -> io::Result<i64>
-    where B: BRefAccess<'a> {
+pub fn parse_piece_index<K, V>(root: &BDictAccess<K, V>) -> io::Result<i64>
+    where V: BRefAccess {
     CONVERT.lookup_and_convert_int(root, PIECE_INDEX_KEY).into()
 }
 
-pub fn parse_total_size<'a, B>(root: &BDictAccess<'a, B>) -> io::Result<i64>
-    where B: BRefAccess<'a> {
+pub fn parse_total_size<K, V>(root: &BDictAccess<K, V>) -> io::Result<i64>
+    where V: BRefAccess {
     CONVERT.lookup_and_convert_int(root, TOTAL_SIZE_KEY).into()
 }
